requirements.txt
openai>=1.0.0
Pillow>=10.0.0
piexif>=1.1.3
google-genai>=1.0.0
google-auth-oauthlib>=1.0.0
requests>=2.28.0
